requirements.txt
python-telegram-bot[job-queue]>=20.0
httpx>=0.24.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
certifi>=2023.7.22
cryptography>=41.0.0
pydantic>=2.0.0